+++ contracts/rust/fairauction/src/contract.rs
@@ -22,7 +22,7 @@ impl FairAuctionFunc {
         FairAuctionFunc { sc: ScContractFunc::new(ctx, HSC_NAME) }
     }
 
-    pub fn delay(&mut self, seconds: i64) -> &mut FairAuctionFunc {
+    pub fn delay(&mut self, seconds: i32) -> &mut FairAuctionFunc {
         self.sc.delay(seconds);
         self
     }
